+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   - `CountLines()` now supports tracing
   - `CountWords()` now supports tracing
   - `CutFields()` now supports tracing
+  - `Dirname()` now supports tracing
   - `DropEmptyLines()` now supports tracing
   - `Echo()` now supports tracing
   - `EchoArgs()` now supports tracing